moved out the tooltip
